requirements.txt
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
docker-pycreds==0.4.0
filelock==3.12.2
flwr==1.4.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
grpcio==1.57.0
huggingface-hub==0.16.4
hydra-core==1.3.2
idna==3.4
iterators==0.0.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numpy==1.25.2
omegaconf==2.3.0
packaging==23.1
pathtools==0.1.2
Pillow==10.0.0
protobuf==3.20.3
psutil==5.9.5
PyYAML==6.0.1
ray==2.6.3
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.9.2
safetensors==0.3.2
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.12
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.31.0
typing_extensions==4.7.1
urllib3==2.0.4
wandb==0.15.8
